+++ wechaty/event.go
@@ -1,4 +1,4 @@
-package impl
+package wechaty
 
 import (
   "github.com/wechaty/go-wechaty/wechaty-puppet/schemas"
