+++ wechaty/interface/wechaty.go
@@ -1,0 +1,5 @@
+package _interface
+
+// Wechaty interface
+type Wechaty interface {
+}
